Add stateless authentication (#49)
add stateless token authentication. The user must be logged in by '/login' by sending a JSON form with user and pass, and in case the authentication is successful, a token is returned in the Header 'X-AUTH-TOKEN', otherwise an authenticated error is sent. Also, when trying to access a resource which is not '/login' or '/' and no token is provided, a 401 unauthorized error will be returned.

- implement login authentication filter
- implement authentication filter for logged users
- implement handler to return unauthorized message
- add codecov config