Limit access to API endpoints that require logged in users to CRUD only their information (#90)
Limit access to API endpoints when trying to do allowed actions but not for your user, but others. If the id of the user sent as the Path Parameter does not match the one in the authentication token, then a 403 Forbidden is returned. It does not make sense to return a 404 Not Found because the user knows the resource exists